pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock_swing"
version = "0.1.0"
description = "Structured trading support and execution-control system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8",
  "PyYAML>=6.0.1",
  "httpx>=0.27",
  "python-dateutil>=2.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
